Move FieldOrderChecker to separate module
