+++ src/eblint/checkers/__init__.py
@@ -1,3 +1,4 @@
 from .base_checker import Checker
 from .violation import Violation
 from .mandatory_field_checker import MandatoryFieldChecker
+from .field_order_checker import FieldOrderChecker
